requirements.txt
lzstring==1.0.4
pycrypto==2.6.1
requests==2.21.0
